% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binning.R
\name{Binning}
\alias{Binning}
\title{Bin spectral data}
\usage{
Binning(data, wl, points_to_bin = 3)
}
\arguments{
\item{data}{Numeric matrix or data frame of spectra (rows = samples, columns = wavelengths)}

\item{wl}{Numeric vector of wavelengths corresponding to columns in data}

\item{points_to_bin}{Integer; number of consecutive points to average per bin}
}
\value{
A list with binned wavelength vector and binned data matrix
}
\description{
Averages every fixed number of wavelength points into bins.
}
