% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Cal_Val.R
\name{Create_Cal_Val}
\alias{Create_Cal_Val}
\title{Short title}
\usage{
Create_Cal_Val(
  data.matrix,
  class.vector,
  wavelength.vector,
  seed,
  feature_selection,
  split_method
)
}
\arguments{
\item{arg}{Description}
}
\value{
What it outputs
}
\description{
What the function does
}
