% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect Artifacts RMSE.R
\name{Detect_Artefacts_RMSE}
\alias{Detect_Artefacts_RMSE}
\title{Short title}
\usage{
Detect_Artefacts_RMSE(parent, child.cor, iqr_factor = 3, min_frac = 0.25)
}
\arguments{
\item{arg}{Description}
}
\value{
What it outputs
}
\description{
What the function does
}
