% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leverage.R
\name{Leverage}
\alias{Leverage}
\title{Short title}
\usage{
Leverage(X, ncomp = 10, frac = 0.15, scale_data = TRUE)
}
\arguments{
\item{arg}{Description}
}
\value{
What it outputs
}
\description{
What the function does
}
