% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorentzian.R
\name{Lorentzian}
\alias{Lorentzian}
\title{Lorentzian Peak Parameter Extraction}
\usage{
Lorentzian(test, wl, window = NULL, minheight)
}
\arguments{
\item{test}{Numeric vector of spectral intensities.}

\item{wl}{Numeric vector of corresponding wavelengths.}

\item{window}{Optional numeric range for peak search.}

\item{minheight}{Minimum peak height threshold.}
}
\value{
Data frame containing PeakID, center (\code{xc}), area (\code{A}), width (\code{w}), asymmetry, and bounds.
}
\description{
Detects peaks and estimates Lorentzian parameters including center, width, area, and asymmetry.
}
