% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monte Carlo.R
\name{Monte_Carlo}
\alias{Monte_Carlo}
\title{Short title}
\usage{
Monte_Carlo(parent, class, child_cor, R = 100, prop_train = 0.5, seed = 123)
}
\arguments{
\item{arg}{Description}
}
\value{
What it outputs
}
\description{
What the function does
}
