% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSID.R
\name{MSID}
\alias{MSID}
\title{Short title}
\usage{
MSID(X, k, metric = "euclidean", max_iter = 50, seed = 123)
}
\arguments{
\item{arg}{Description}
}
\value{
What it outputs
}
\description{
What the function does
}
