% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Correction.R
\name{plot_pca_correction}
\alias{plot_pca_correction}
\title{Visualize PCA Correction Results}
\usage{
plot_pca_correction(pca_parent, pca_child, result, dims = c(1, 2))
}
\arguments{
\item{pca_parent}{Parent PCA object}

\item{pca_child}{Child PCA object}

\item{result}{Output from pca_tps_correct}

\item{dims}{Dimensions to plot (default = c(1,2))}
}
\description{
Creates diagnostic plots for PCA correction in 2D or 3D.
}
