% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Cal_Val_PDS.R
\name{Create_Cal_Val_PDS}
\alias{Create_Cal_Val_PDS}
\title{Short title}
\usage{
Create_Cal_Val_PDS(
  parent,
  child,
  class.vector,
  wl.vector,
  seed,
  feat_select = NA,
  split_method = "partition",
  p = 0.1,
  k = 5,
  remove_classes = NULL,
  ncomp_svd = NULL
)
}
\arguments{
\item{arg}{Description}
}
\value{
What it outputs
}
\description{
What the function does
}
